feat: abstract table model, add table filters column
